+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Fixed
+- Displaying users kyc details(general)
+
 ## [1.0.0] - 2019-08-15
 
 [Unreleased]: https://github.com/tokend/admin-panel/compare/1.0.0...HEAD